+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 I will record changes to this file just so i don't need to look at commit history every time
 
 # Changelog
+2025-10-07:
+- Added setup/debian_bootstrap.sh script for initial Debian setup
 2025-10-04:
 - Added helm to the devcontainer. Confirmed by installing monitoring stack with helm
 - Added setup/SETUP.md with the setup plan and a diagram in mermaid syntax